aviara-human/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1
httpx==0.28.0
anthropic==0.40.0
redis==5.2.1
celery[redis]==5.4.0

# dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
